arm: zynq: xilinx_axidma: fix race condition in xilinx_chan_desc_cleanup
xilinx_chan_desc_cleanup uses a spinlock to prevent multiple threads
from accessing the same descriptor. The problem arrises when there
are at least two descriptors in active_list which require callback
and are both done because before callback, the spinlock is released.
Suppose the following:
- Thread 1 acquires the lock, enters the loop with desc pointing
to A and _desc pointing to B. Deletes A from the list, releases the lock,
and makes the callback.
- Thread 2 acquires the lock, enters the loop with desc pointing
to B and _desc pointing to next element. Deletes B from the list,
releases the lock.
- Thread 1 reacquires the lock, does the next iteration of the loop with
desc pointing to B (which was deleted by Thread 2) - race condition.

The proposed solution is to iterate while there are elements in the list,
get the head descriptor and remove it from list before callback.

Signed-off-by: Nicolae Rosia <nicolae.rosia@certsign.ro>
Reported-by: Joseph Zavodny <joseph.zavodny@jhuapl.edu>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 2e0eef9126e87ad9371f169f49825f415875641a)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>