ARM: zynq: xilinx_axidma: get rid of useless 'feature' variable
The structures xilinx_dma_chan and xilinx_dma_device have a variable
which encode 2 capabilities:
 - device has Scatter Gatter
 - Control Stream and Status Stream.
For Scatter Gatter, there already was a variable for this, called has_sg
and Control Stream and Status Stream is not being used.
With this ocassion, we also get rid of the defines which encode capabilities.

Signed-off-by: Nicolae Rosia <nicolae.rosia@certsign.ro>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit c239960540d83cd02fa075c587aa937fcaaad89e)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>