serial: 8250_pci: mask UART TX completion intr in byt_set_termios
The code masking for TX completion interrupt was added as part of
enabling Intel Baytrail UART. At that time, this code executes
during initialization. Regression test on power management
suspend/resume reveals the need for this bit to be masked again
during controller resume. So, instead of doing the mask during
initialization, put it in byt_set_termios to make sure that the
bit is properly set even on the event of system suspend/resume.

Signed-off-by: Maurice Petallo <mauricex.r.petallo@intel.com>