mips/kvm: Rename kvm_vcpu_arch.pc to kvm_vcpu_arch.epc
From: David Daney <david.daney@cavium.com>

The proper MIPS name for this register is EPC, so use that.
Change the asm-offsets name to KVM_VCPU_ARCH_EPC, so that the symbol
name prefix matches the structure name.

Rename VCPU_registername to KVM_VCPU_ARCH_registername
This makes it follow the pattern where the structure name is the
symbol name prefix.

Signed-off-by: David Daney <david.daney@cavium.com>
Acked-by: Ralf Baechle <ralf@linux-mips.org>
Signed-off-by: Abhishek Paliwal <abhishek.paliwal@aricent.com>