i2c: allow Designware I2C to be probed before SMBus for Baytrail
When SMBus is enabled in Baytrail, it is being probed first before
all the Designware I2C devices. This means SMBus will take the first
available bus id which is 0 and then the I2C devices will take the
next succeeding bus ids. This should be fine as long as the I2C
devices bus ids are assigned dynamically.

The problem arises when the I2C devices' bus ids are assigned
statically and the SMBus is probed first. The kernel will fail to
add the i2c adapter because something else, which in that case is
the SMBus, has taken the bus id that the i2c adapter is trying to
take.

As mentioned in commit "i2c: designware-pci: use static bus num
allocation for baytrail", it is required for I2C devices' bus ids
to be assigned statically for LPE Audio to register the i2c properly.

To resolve this, we either force the SMBus to take the next bus id
after all the designware I2C devices or we allow all the I2C devices
to be probed first before the SMBus. This patch implements the 2nd.

Signed-off-by: Maurice Petallo <mauricex.r.petallo@intel.com>