serial: 8250_core: handle_irq returns 1 only if data was processed
There are cases when the serial8250 interrupt routine is being called
but there are really no data to process (IIR bit 3:0 is 0x1, UART_IIR_NO_INT).
The good thing is, the routine just ignores it and returns back
immediately.

There are also cases when the routine handles only rx-related interrupts.
When FIFO is enabled and THRE is disabled, LSR_THRE (LSR bit 5) will
indicate that TX FIFO is empty which is correct since the port, in this
case, is configured only for receiving. But, in the driver code, this will
cause the function serial8250_handle_irq() to call serial8250_tx_chars()
which only ends up returning from:

if (uart_circ_empty(xmit))
{
	__stop_tx(up); return;
}

serial8250_handle_irq() returns 1 and the serial8250_interrupt() will then
continue executing the do-while loop which keeps on trying and trying to
process data even when there's actually nothing to transfer. This will
result to serial8250 driver saying:

"serial8250: too much work for irq..."

This patch will allow serial8250_handle_irq() to return 1 only when there's
actually some data that was processed. Otherwise, the routine ignores it just
like how UART_IIR_NO_INT is being handled. Also, we add an extra check before
calling serial8250_tx_chars() inside serial8250_handle_irq() to see
whether there are data to be transmitted and the tx is not stopped.

Signed-off-by: Maurice Petallo <mauricex.r.petallo@intel.com>
Signed-off-by: Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>