MIPS: Expand __swp_offset() to carry 32 significant bits of data.
From: David Daney <david.daney@cavium.com>

With CONFIG_MIGRATION, the PFN of the migrating pages is stored in
__swp_offset(), so we must have enough bits to store the largest
possible PFN.  OCTEON NUMA systems have 41 bits of physical address
space, so with 4K pages (12-bits), we need at least 29 bits to store
the PFN.

The current width of 24-bits is too narrow, so expand it to 32-bits.
This leaves the low order 24 bits as zero which will continue to not
interfere with the PTE 'G' and 'V' bits.

Signed-off-by: David Daney <david.daney@cavium.com>
Signed-off-by: Abhishek Paliwal <abhishekpaliwal@aricent.com>