spi/pxa2xx: auto switch between PIO and DMA with configurable threshold size
This commit introduces a mode where spidev driver can enable/disable
an automatic switching for data transfer to be performed in PIO mode or
DMA mode upon reaching a configurable data transfer threshold size.

If the data transfer size larger than the threshold size and less than the
value of MAX_DMA_LEN, the transfer is performed in DMA mode.

Signed-off-by: Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>