mmc: block: refactor read-only handling
There are several reasons why gendisk corresponding to SD/eMMC card or
hardware partition is read-only for the system:
- underlying device is readonly (due to write protect GPIO, or due to
  card lacks support for write commands),
- due to card-level write protection (currently supported only for eMMC
  boot partitions),
- due to software 'force_ro' attribute (currently enabled by default for
  eMMC boot partitions).

This is managed indirectly, via gendisk's read-only flag.

Since more possible read-only reasons are about to be added, need to
replace that with explicit handling.

This patch turns 'read_only' field of struct mmc_blk_data into a bitmask,
with individual bits representing possible read-only reasons. Gendisk
is marked read-only when at least one bit is set in the mask, and
is marked read-write when last bit is cleared.

Signed-off-by: Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>