aufs3: mmap: Fix races in madvise_remove() and sys_msync()
In madvise_remove() and sys_msync() we drop the mmap_sem before
dropping references to the mapped file(s).  As soon as we drop the
mmap_sem, the vma we got them from might be destroyed by another
thread, so calling vma_do_fput() is a possible use-after-free.

In these cases we don't actually need a reference to the aufs file, so
revert to using get_file() and fput() directly.

Fixes CVE-2015-7312.
Upstream-Status: Backport

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: Sona Sarmadi <sona.sarmadi@enea.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>